cleanup some compilation warnings

+++ drivers/power/ab8500_fg.c
@@ -1732,7 +1732,7 @@ static void ab8500_fg_charge_state_to(struct ab8500_fg *di,
 static void ab8500_fg_discharge_state_to(struct ab8500_fg *di,
 	enum ab8500_fg_charge_state new_state)
 {
-	if (di->discharge_state == new_state)
+	if ((int)di->discharge_state == (int)new_state)
 		return;
 
 	dev_dbg(di->dev, "Discharge state from %d [%s] to %d [%s]\n",
